% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sops.R
\name{lp_to_probs}
\alias{lp_to_probs}
\title{Convert Cumulative Log-Odds to Probabilities (Internal)}
\usage{
lp_to_probs(eta, M)
}
\arguments{
\item{eta}{Matrix of linear predictors with dimensions N x M (N observations by M thresholds)}

\item{M}{Number of thresholds (one less than the number of categories)}
}
\value{
Matrix of probabilities with dimensions N x (M+1)
}
\description{
Converts a matrix of cumulative log-odds to category probabilities.
}
\keyword{internal}
