% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz.R
\name{plot_results}
\alias{plot_results}
\title{ggplot wrapper to present simulation results}
\usage{
plot_results(data, ..., x, group, ggplot_options = list(), combine = TRUE)
}
\arguments{
\item{data}{A data frame of results.}

\item{...}{Operating characteristics to be displayed on the y-axis (in different plots)}

\item{x}{What to plot on the x axis.}

\item{group}{Typically different analysis strategies}

\item{ggplot_options}{List of further options passed to ggplot}

\item{combine}{Should the plots for different operating characteristics}
}
\value{
A ggplot list and optionally a patchwork object
}
\description{
A result data frame with operating characteristic estimates is plotted in a
unified way.
}
\details{
The input is typically a data frame where in each row the operating characteristics
of a strategy in a certain scenario are contained.

... specifies what to plot on the y-axis, which can be different inputs -
typically operating characteristics - and a separate plot for each input is generated.

x determines the column of values to be plotted on the x-axis.

group will typically be the analysis strategies compared and will be lines in
different colours. When combined = TRUE, the single plots and a patchwork::wrap_plots
with all plots is returned.

ggplot_options is integrated into a list of ggplot options and directly passed
to the ggplot calls.
}
\examples{
\dontrun{
...
}


}
