% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sops.R
\name{sop_sim}
\alias{sop_sim}
\title{Simulate state occupancy probabilities}
\usage{
sop_sim(
  model,
  baseline_data,
  varname = "y",
  pvarname = "yprev",
  tvarname = "time",
  times = 1:30,
  ylevels = factor(1:6, ordered = TRUE),
  absorb = "6",
  n_sim = 1000
)
}
\description{
Simulate state occupancy probabilities
}
