% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taooh_bootstrap.R
\name{taooh_bootstrap}
\alias{taooh_bootstrap}
\title{Use bootstrap to compute confidence intervals for the time alive and out of
hospital}
\usage{
taooh_bootstrap(
  data,
  n_boot,
  formula,
  workers = NULL,
  parallel = TRUE,
  ylevels = 1:6,
  absorb = 6,
  times = 1:max(data[["time"]]),
  varnames = list(tvarname = "time", pvarname = "yprev", id = "id", tx = "tx")
)
}
\arguments{
\item{data}{A data frame containing the patient data}

\item{n_boot}{Number of bootstrap samples}

\item{formula}{Model formula for orm}

\item{workers}{Number of workers used for parallelization. Default is availableCores()-1}

\item{parallel}{Whether parallelization should be used}

\item{ylevels}{States in the data}

\item{absorb}{Absorbing state}

\item{times}{Time points in the data. Default is the maximum time observed in the data}

\item{varnames}{List of variable names in the data}
}
\description{
Details...
}
\details{
Output is the sample estimate of time alive and out of hospital, and an upper
and lower confidence bound for the estimate.
}
\examples{
\dontrun{
taooh_bootstraps()
}
}
\keyword{alive}
\keyword{and}
\keyword{bootstrap}
\keyword{hospital}
\keyword{of}
\keyword{out}
\keyword{time}
