% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vglm_helpers.R
\name{vglm.markov}
\alias{vglm.markov}
\title{Fit a VGAM Markov Model}
\usage{
vglm.markov(
  formula,
  family = stop("argument 'family' needs to be assigned"),
  data = list(),
  weights = NULL,
  subset = NULL,
  na.action,
  etastart = NULL,
  mustart = NULL,
  coefstart = NULL,
  control = VGAM::vglm.control(...),
  offset = NULL,
  model = FALSE,
  x.arg = TRUE,
  y.arg = TRUE,
  contrasts = NULL,
  constraints = NULL,
  extra = list(),
  form2 = NULL,
  qr.arg = TRUE,
  smart = TRUE,
  ...
)
}
\arguments{
\item{formula}{
  a symbolic description of the model to be fit.
  The RHS of the formula is applied to each linear
  predictor.
  The effect of different variables in each linear predictor
  can be controlled by specifying constraint matrices---see
  \code{constraints} below.


  }

\item{family}{A VGAM family object, e.g. \code{VGAM::cumulative()}.}

\item{data}{
  an optional data frame containing the variables in the model.
  By default the variables are taken from
  \code{environment(formula)}, typically the environment
  from which \code{vglm} is called.




  }

\item{weights}{
  an optional vector or matrix of (prior fixed and known) weights
  to be used in the fitting process.
  If the \pkg{VGAM} family function handles multiple responses
  (\eqn{Q > 1} of them, say) then
  \code{weights} can be a matrix with \eqn{Q} columns.
  Each column matches the respective response.
  If it is a vector (the usually case) then it is recycled into a
  matrix with \eqn{Q} columns.
  The values of \code{weights} must be positive; try setting
  a very small value such as \code{1.0e-8} to effectively
  delete an observation.



% 20201215:

  Currently the \code{weights} argument supports sampling
  weights from complex sampling designs
  via \pkg{svyVGAM}.
  Some details can be found at
  \url{https://CRAN.R-project.org/package=svyVGAM}.



% Creates CRAN problems:
% \url{cran.r-project.org/web/packages/
%      svyVGAM/vignettes/theory.html}.





% 20140507:
% Currently the \code{weights} argument does not support
% sampling weights from complex sampling designs.
% And currently sandwich estimators are not computed in any
% shape or form.
% The present weights are multiplied by the corresponding
% log-likelihood contributions and added to form the
% overall log-likelihood.






% If \code{weights} is a matrix,
% then it should be must be in \emph{matrix-band} form,
% whereby the
% first \eqn{M} columns of the matrix are the diagonals,
% followed by the upper-diagonal band, followed by the
% band above that, etc. In this case, there can be up to
% \eqn{M(M+1)} columns, with the last column corresponding
% to the (1,\eqn{M}) elements of the weight matrices.


  }

\item{subset}{
  an optional logical vector specifying a subset of
  observations to
  be used in the fitting process.



  }

\item{na.action}{
  a function which indicates what should happen when
  the data contain \code{NA}s.
  The default is set by the \code{na.action} setting
  of \code{\link[base]{options}},
  and is \code{na.fail} if that is unset.
  The ``factory-fresh'' default is \code{\link[stats]{na.omit}}
  which is known as \emph{complete case analysis}
  and applied to both sides of the formula.



% 20240308; changed na.action = na.fail to just na.action.



  }

\item{etastart}{
  optional starting values for the linear predictors.
  It is a \eqn{M}-column matrix with the same
  number of rows as
  the response.
  If \eqn{M = 1} then it may be a vector.
  Note that \code{etastart} and the output
  of \code{predict(fit)}
  should be comparable.
  Here, \code{fit} is the fitted object.
  Almost all \pkg{VGAM} family functions are self-starting.
  
  

  }

\item{mustart}{
  optional starting values for the fitted values.
  It can be a vector or a matrix;
  if a matrix, then it has the same number of rows
  as the response.
  Usually \code{mustart} and the output of \code{fitted(fit)}
  should be comparable.
  Most family functions do not make use of this argument
  because it is not possible to compute all \eqn{M} columns of
  \code{eta} from \code{mu}.
  
  

  }

\item{coefstart}{
  optional starting values for the coefficient vector.
  The length and order must match that of \code{coef(fit)}.


  }

\item{control}{
  a list of parameters for controlling the fitting process.
  See \code{\link[VGAM]{vglm.control}} for details.


  }

\item{offset}{
   a vector or \eqn{M}-column matrix of offset values.
   These are \emph{a priori} known and are added to the
   linear/additive predictors during fitting.


  }

\item{model}{
  a logical value indicating whether the
  \emph{model frame}
  should be assigned in the \code{model} slot.


  }

\item{x.arg, y.arg}{
  logical values indicating whether
  the LM matrix and response vector/matrix used in the fitting
  process should be assigned in the \code{x} and \code{y} slots.
  Note that the model matrix is the LM matrix; to get the VGLM
  matrix type \code{model.matrix(vglmfit)} where
  \code{vglmfit} is a \code{vglm} object.


  }

\item{contrasts}{
  an optional list. See the \code{contrasts.arg}
  of \code{\link{model.matrix.default}}.


  }

\item{constraints}{Optional VGAM constraints list. For inline \code{rcs()} terms,
names should match the column-level constraint names in a full proportional
odds fit returned by \code{vglm.markov()}.}

\item{extra}{
    an optional list with any extra information that
    might be needed by
  the \pkg{VGAM} family function.


  }

\item{form2}{
  the second (optional) formula.
  If argument \code{xij} is used
  (see \code{\link[VGAM]{vglm.control}}) then
  \code{form2} needs to have \emph{all} terms in the model.
  Also, some \pkg{VGAM} family functions
  such as \code{\link[VGAM]{micmen}}
  use this argument to input the regressor variable.
  If given, the slots \code{@Xm2} and \code{@Ym2}
  may be assigned.
  Note that smart prediction applies to terms
  in \code{form2} too.


  }

\item{qr.arg}{
  logical value indicating whether the slot \code{qr}, which
  returns the QR decomposition of the VLM model matrix,
  is returned on the object.


  }

\item{smart}{
  logical value indicating whether smart prediction
  (\code{\link[VGAM]{smartpred}}) will be used.


  }

\item{...}{
  further arguments passed into \code{\link[VGAM]{vglm.control}}.


  }
}
\value{
A fitted S4 \code{vglm} object with an internal Markov marker attribute.
}
\description{
\code{vglm.markov()} is the recommended VGAM fitting entrypoint for
\code{markov.misc} SOP workflows. It follows \code{VGAM::vglm()} closely and returns a
normal S4 \code{vglm} object, but marks the fit so the SOP prediction code can use
package-native Markov prediction. For inline restricted cubic spline terms
such as \code{rcs(time, 4)} or \code{rms::rcs(time, 4)}, the internal VGAM assignment
metadata is split by generated spline column. This makes it possible to give
individual spline basis columns separate constraint matrices for partial
proportional odds models.
}
\details{
Non-spline terms, factor terms, and workflows that use explicit precomputed
basis columns keep ordinary VGAM assignment behavior. This function does not
compute robust covariance matrices; wrap the returned fit with
\code{\link[=robcov_vglm]{robcov_vglm()}} when robust covariance is needed.
}
\examples{
\dontrun{
library(VGAM)
library(rms)

fit_po <- vglm.markov(
  ordered(y) ~ rcs(time, 4) + tx + age + yprev,
  family = cumulative(reverse = TRUE, parallel = TRUE),
  data = data
)

constraints <- fit_po@constraints
linear_time_term <- grep("^rcs\\\\(time, 4\\\\).*time$", names(constraints),
  value = TRUE
)[[1]]
n_thresholds <- nrow(constraints[[linear_time_term]])
constraints[[linear_time_term]] <- cbind(
  PO_effect = rep(1, n_thresholds),
  linear_deviation = seq_len(n_thresholds) - 1
)

fit_ppo <- vglm.markov(
  ordered(y) ~ rcs(time, 4) + tx + age + yprev,
  family = cumulative(reverse = TRUE, parallel = FALSE),
  data = data,
  constraints = constraints
)
}

}
\seealso{
\code{\link[=robcov_vglm]{robcov_vglm()}}, \code{\link[=avg_sops]{avg_sops()}}, \code{\link[=sops]{sops()}}
}
