% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sops.R
\name{array_to_df_individual}
\alias{array_to_df_individual}
\title{Convert SOPs Array to Individual-Level Data Frame}
\usage{
array_to_df_individual(sops_array, times, ylevels, newdata, by = NULL)
}
\arguments{
\item{sops_array}{Array of dimensions (n_pat x n_times x n_states).}

\item{times}{Vector of time points.}

\item{ylevels}{Vector of state levels.}

\item{newdata}{Original data with rowid.}

\item{by}{Optional character vector of variables to aggregate by.}
}
\value{
Data frame with individual SOPs (or aggregated if by is specified).
}
\description{
Convert SOPs Array to Individual-Level Data Frame
}
\keyword{internal}
