% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sops_new.R
\name{array_to_df_individual}
\alias{array_to_df_individual}
\title{Convert SOPs Array to Individual-Level Data Frame}
\usage{
array_to_df_individual(sops_array, times, ylevels, newdata)
}
\arguments{
\item{sops_array}{Array of dimensions (n_pat x n_times x n_states).}

\item{times}{Vector of time points.}

\item{ylevels}{Vector of state levels.}

\item{newdata}{Original data with rowid.}
}
\value{
Data frame with individual SOPs.
}
\description{
Convert SOPs Array to Individual-Level Data Frame
}
\keyword{internal}
