% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robcov_vglm.R
\name{coef.robcov_vglm}
\alias{coef.robcov_vglm}
\title{Extract Coefficients from robcov_vglm Objects}
\usage{
\method{coef}{robcov_vglm}(object, ...)
}
\arguments{
\item{object}{A robcov_vglm object.}

\item{...}{Additional arguments (ignored).}
}
\value{
Named numeric vector of coefficients.
}
\description{
Extract Coefficients from robcov_vglm Objects
}
