% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sops_new.R
\name{marginalize_sops_array}
\alias{marginalize_sops_array}
\title{Marginalize SOPs Array Over Patients}
\usage{
marginalize_sops_array(
  sops_array,
  grid,
  times,
  ylevels,
  variables,
  n_cf,
  n_each
)
}
\arguments{
\item{sops_array}{Array of dimensions (n_pat x n_times x n_states).}

\item{grid}{Data frame of variable combinations.}

\item{times}{Vector of time points.}

\item{ylevels}{Vector of state levels.}

\item{variables}{Named list of variables.}

\item{n_cf}{Number of counterfactual scenarios.}

\item{n_each}{Number of patients per scenario.}
}
\value{
Data frame with marginalized SOPs.
}
\description{
Averages individual-level SOPs to get population-average (marginal) SOPs.
}
\keyword{internal}
