% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markov-misc-package.R
\docType{package}
\name{markov.misc-package}
\alias{markov.misc}
\alias{markov.misc-package}
\title{markov.misc native routines}
\description{
Estimate state occupancy probabilities and treatment contrasts for longitudinal ordinal outcomes using discrete-time Markov transition models. Supports (partial) proportional-odds model backends from 'VGAM', 'rms', and 'rmsb'.
}
\author{
\strong{Maintainer}: Johannes Schwenke \email{johannes.schwenke@gmail.com} [copyright holder]

Authors:
\itemize{
  \item Felix Herkner \email{felix.herkner@meduniwien.ac.at} [copyright holder]
}

}
\keyword{internal}
