% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robcov_vglm.R
\name{vcov.robcov_vglm}
\alias{vcov.robcov_vglm}
\title{Extract Variance-Covariance Matrix from robcov_vglm Objects}
\usage{
\method{vcov}{robcov_vglm}(object, ...)
}
\arguments{
\item{object}{A robcov_vglm object.}

\item{...}{Additional arguments (ignored).}
}
\value{
The robust variance-covariance matrix.
}
\description{
Returns the robust (sandwich) variance-covariance matrix.
}
